% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_stype.R
\name{reg_stype}
\alias{reg_stype}
\title{Fit a model using S-type robust estimator}
\usage{
reg_stype(y, x, c)
}
\arguments{
\item{y}{A numeric vector of responses.}

\item{x}{A numeric vector of predictors.}
}
\value{
A list containing the model coefficients and diagnostics.
}
\description{
This function fits a regression model using the S-type estimator.
}
