% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regstype.R
\name{regstype}
\alias{regstype}
\title{Fit a regression model using the S-type estimators.
(Fully Optimized Version)}
\usage{
regstype(y, x)
}
\arguments{
\item{y}{Dependent variables (Dataframe, matrix, vector).}

\item{x}{Explanatory variables (Dataframe, matrix, vector).}
}
\value{
A list containing the model coefficients and diagnostics.
}
\description{
Fit a regression model using the S-type estimators.
(Fully Optimized Version)
}
