% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regweighteds.R
\name{regweighteds}
\alias{regweighteds}
\title{Weighted regression analysis}
\usage{
regweighteds(y, x, W)
}
\arguments{
\item{y}{Dependent variables (Dataframe, vector).}
\item{x}{Explanatory variables (Dataframe, matrix).}
\item{W}{A numeric vector of weights.}
}
\value{
A list containing the regression model results.
}
\description{
This function performs weighted regression analysis.
}
