% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regstype.R
\name{regstype}
\alias{regstype}
\title{Fit a regression model using the S-type estimators}
\usage{
regstype(y, x)
}
\arguments{
\item{y}{Dependent variables (Dataframe, vector).}

\item{x}{Explanatory variables (Dataframe, matrix).}
}
\value{
A list containing the model coefficients and diagnostics.
}
\description{
This function fits a regression model using the S-type estimators.
}
