% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regweighteds.R
\name{regweighteds}
\alias{regweighteds}
\title{Weighted regression commands}
\usage{
regweighteds(y, x, W)
}
\arguments{
\item{y}{A numeric vector of responses.}

\item{weights}{A numeric vector of weights.}
}
\value{
A list containing the model results.
}
\description{
This function performs weighted regression analysis using stype robust estimators.
}
